tests has been fixed according to mentor advice

+++ src/test/java/service/AmountHandlerPurchaseTest.java
@@ -8,6 +8,6 @@ public class AmountHandlerPurchaseTest {
     @Test
     public void getAmount() {
         AmountHandlerPurchase amountHandlerPurchase = new AmountHandlerPurchase();
-        Assert.assertEquals("",(int)-10,(int)amountHandlerPurchase.getAmount(10));
+        Assert.assertEquals("",Integer.valueOf(-10),amountHandlerPurchase.getAmount(10));
     }
 }
