+++ src/test/java/service/AmountHandlerReturnTest.java
@@ -8,6 +8,6 @@ public class AmountHandlerReturnTest {
     @Test
     public void getAmount() {
         AmountHandlerReturn amountHandlerReturn = new AmountHandlerReturn();
-        Assert.assertEquals("",(int)10,(int)amountHandlerReturn.getAmount(10));
+        Assert.assertEquals("",Integer.valueOf(10),amountHandlerReturn.getAmount(10));
     }
 }
