+++ src/test/java/service/AmountHandlerSupplyTest.java
@@ -8,6 +8,6 @@ public class AmountHandlerSupplyTest {
     @Test
     public void getAmount() {
         AmountHandlerSupply amountHandlerSupply = new AmountHandlerSupply();
-        Assert.assertEquals("",(int)10,(int)amountHandlerSupply.getAmount(10));
+        Assert.assertEquals("",Integer.valueOf(10),amountHandlerSupply.getAmount(10));
     }
 }
